Possible fix for https://github.com/square/picasso/issues/769
When Center Crop is used, a single scale factor is calculated to transform the input image to a target image size. This does not work for all input cases. I'm not sure if this is the best solution but hope this at least starts a discussion. The summary of the problem: integers multiplied by a transformation matrix of floats cannot always create the requested targetSize. A single scale factor cannot be used for both X and Y because the input sizes are integers. Test included.