Fixing Message memory leaks
Prior to Android L, the VM caches stack local variables in blocking loops. This can lead to memory leaks of messages for threads that become inactive but aren't shutdown. This change ensure that we never leak for too long and always replace the leaking reference.

These memory leaks tend to be problematic when using Picasso in an app that also uses Dialogs, because Dialogs obtain template Message instance that are never recycled. These messages are listeners and often hold references to the activity. When the dialog is GCed, the messages should be GCed too but they aren't because of the thread stack local message leak.